Handle corner case of zero-VTE AIS track(s)
It looks like there is a small chance of getting one or more AIS tracks that
graze the selection bounding box but have zero VTE in the bbox.
That probably only happens after we have expanded the bbox to mitigate GeoTIFF
smearing of VTEs.
This change resolves that failure mode by choosing a track without using VTE
weighting (i.e. with uniform probability).