Docs maintenance (#73)
* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures compatibility with
the latest environment settings available on readthedocs.

* Update Sphinx & optional dependencies version pins

Updated Sphinx and optional dependencies packages pinned versions in
`environment-dev.yaml`, `environment-test.yaml`, and `environment-rtd.yaml`
in preparation for Python 3.13 environment.

* Update Sphinx source_suffix config to mapping

The deprecated list of strings appears to have been removed in Sphinx=8.

* Improve grammar & markup in docstring

* Update redirected links found by linkcheck

* Update docs example output from Sphinx

Updated Sphinx build and linkcheck commands example outputs.