Update to Python 3.13 (#80)
* Add Python 3.13 to GitHub pytest workflow matrices

Updated the test matrices in `pytest-with-coverage.yaml` and
`pytest-no-midoss .yaml` to include Python 3.13.

* Update latest supported Python version to 3.13

Upgraded Python from 3.12 to 3.13 across all environment files to ensure
compatibility with the latest features and fixes.

* Update Sphinx linkcheck workflow to Python 3.13

Modified the Python version in the matrix configuration from 3.12 to 3.13 in
`sphinx-linkcheck.yaml`. This ensures compatibility testing with the latest
Python version.

* Update to Python 3.13 for documentation and use

Changed Python version references from 3.12 to 3.13 in README and development
documentation to align with the latest Python development environment.
Updated example pytest and Sphinx linkcheck output in dev docs accordingly.