Fix GHA `pytest-no-midoss` workflow (#108)
* Use matrix.python-version in GHA pytest workflow

Replaced `inputs.python-version` with `matrix.python-version` in the
`pytest-no-midoss` workflow. This fixes a copy-paste error from the reusable
`pytest-with-coverage` workflow where the Python version is an input.

* Add date-based cache key to GHA pytest workflow

Introduced a missing step to generate a date-based cache key for downloads in
the `pytest-no-midoss` workflow. This ensures that downloads are cached for 1
day. It is unclear how long the downloads were cached previously when the date
part of the cache key was empty.